requirements.txt
pillow==11.0.0
cairosvg==2.7.1
